Cleanup and test fixes for MasterPlaylistController Add eslint ignores for TODOs in master-playlist-controller and videojs-contrib-hls. Fix linting error in playlist-loader. Move some generic test functions from videojs-contrib-hls' tests to plugin-helpers. Add some master-playlist-controller tests. Add excludeIncompatibleVariants logic back in on loadedmetadata. Move some relevant tests over from videojs-contrib-hls' tests to master-playlist-controller's tests. Add TODOs for skipped source buffer related tests, as responsibility for source buffer has shifted to segment loaders.
