+++ test/master-playlist-controller.test.js
@@ -1,0 +1,388 @@
+import QUnit from 'qunit';
+import videojs from 'video.js';
+import {
+  useFakeEnvironment,
+  useFakeMediaSource,
+  createPlayer,
+  standardXHRResponse,
+  openMediaSource
+} from './plugin-helpers.js';
+import MasterPlaylistController from '../src/master-playlist-controller';
+/* eslint-disable no-unused-vars */
+// we need this so that it can register hls with videojs
+import { Hls } from '../src/videojs-contrib-hls';
+/* eslint-enable no-unused-vars */
+
+QUnit.module('MasterPlaylistController', {
+  beforeEach() {
+    this.env = useFakeEnvironment();
+    this.clock = this.env.clock;
+    this.requests = this.env.requests;
+    this.mse = useFakeMediaSource();
+
+    // force the HLS tech to run
+    this.origSupportsNativeHls = videojs.Hls.supportsNativeHls;
+    videojs.Hls.supportsNativeHls = false;
+
+    this.player = createPlayer();
+    this.player.src({
+      src: 'manifest/master.m3u8',
+      type: 'application/vnd.apple.mpegurl'
+    });
+    this.masterPlaylistController = this.player.tech_.hls.masterPlaylistController_;
+  },
+  afterEach() {
+    this.env.restore();
+    this.mse.restore();
+    videojs.Hls.supportsNativeHls = this.origSupportsNativeHls;
+  }
+});
+
+QUnit.test('throws error when given an empty URL', function() {
+  let options = {
+    url: 'test',
+    currentTimeFunc: () => {},
+    hlsHandler: this.masterPlaylistController.hlsHandler
+  };
+
+  QUnit.ok(new MasterPlaylistController(options), 'can create with options');
+
+  options.url = '';
+  QUnit.throws(() => {
+    new MasterPlaylistController(options); // eslint-disable-line no-new
+  }, /A non-empty playlist URL is required/, 'requires a non empty url');
+});
+
+QUnit.test('obeys preload option', function() {
+  this.player.preload('none');
+  // master
+  standardXHRResponse(this.requests.shift());
+  // playlist
+  standardXHRResponse(this.requests.shift());
+
+  QUnit.equal(this.requests.length, 0, 'no segment requests when preload is none');
+
+  this.player.src({
+    src: 'manifest/master.m3u8',
+    type: 'application/vnd.apple.mpegurl'
+  });
+  this.masterPlaylistController = this.player.tech_.hls.masterPlaylistController_;
+
+  this.player.preload('auto');
+  // master
+  standardXHRResponse(this.requests.shift());
+  // playlist
+  standardXHRResponse(this.requests.shift());
+
+  QUnit.equal(this.requests.length, 1, 'segment request when preload is auto');
+});
+
+QUnit.test('tech fires loadedmetadata when playlist loader loads first playlist',
+function() {
+  let firedLoadedMetadata = false;
+
+  this.masterPlaylistController.hlsHandler.tech_.on('loadedmetadata', () => {
+    firedLoadedMetadata = true;
+  });
+
+  // master
+  standardXHRResponse(this.requests.shift());
+  QUnit.ok(!firedLoadedMetadata, 'did not fire loadedmetadata');
+  // playlist
+  standardXHRResponse(this.requests.shift());
+  QUnit.ok(firedLoadedMetadata, 'fired loadedmetadata');
+});
+
+QUnit.test('creates combined and audio only SegmentLoaders', function() {
+  QUnit.equal(this.masterPlaylistController.mainSegmentLoader_.state, 'INIT',
+              'created combined segment loader');
+  QUnit.equal(this.masterPlaylistController.audioSegmentLoader_.state, 'INIT',
+              'created alternate audio track segment loader');
+});
+
+QUnit.test('if buffered, will request second segment byte range', function() {
+  this.requests.length = 0;
+  this.player.src({
+    src: 'manifest/playlist.m3u8',
+    type: 'application/vnd.apple.mpegurl'
+  });
+  this.masterPlaylistController = this.player.tech_.hls.masterPlaylistController_;
+
+  this.player.tech_.triggerReady();
+  this.clock.tick(1);
+  this.player.tech_.trigger('play');
+  openMediaSource(this.player, this.clock);
+  this.masterPlaylistController.mainSegmentLoader_.sourceUpdater_.buffered = () => {
+    return videojs.createTimeRanges([[0, 20]]);
+  };
+  // playlist
+  standardXHRResponse(this.requests[0]);
+  // segment
+  standardXHRResponse(this.requests[1]);
+  this.masterPlaylistController.mediaSource.sourceBuffers[0].trigger('updateend');
+  this.clock.tick(10 * 1000);
+  QUnit.equal(this.requests[2].headers.Range, 'bytes=1823412-2299991');
+});
+
+QUnit.test('re-initializes the combined playlist loader when switching sources',
+function() {
+  openMediaSource(this.player, this.clock);
+  // master
+  standardXHRResponse(this.requests.shift());
+  // playlist
+  standardXHRResponse(this.requests.shift());
+  // segment
+  standardXHRResponse(this.requests.shift());
+  // change the source
+  this.player.src({
+    src: 'manifest/master.m3u8',
+    type: 'application/vnd.apple.mpegurl'
+  });
+  this.masterPlaylistController = this.player.tech_.hls.masterPlaylistController_;
+  // maybe not needed if https://github.com/videojs/video.js/issues/2326 gets fixed
+  this.clock.tick(1);
+  QUnit.ok(!this.masterPlaylistController.masterPlaylistLoader_.media(),
+           'no media playlist');
+  QUnit.equal(this.masterPlaylistController.masterPlaylistLoader_.state,
+              'HAVE_NOTHING',
+              'reset the playlist loader state');
+  QUnit.equal(this.requests.length, 1, 'requested the new src');
+
+  // buffer check
+  this.clock.tick(10 * 1000);
+  QUnit.equal(this.requests.length, 1, 'did not request a stale segment');
+
+  // sourceopen
+  openMediaSource(this.player, this.clock);
+
+  QUnit.equal(this.requests.length, 1, 'made one request');
+  QUnit.ok(
+    this.requests[0].url.indexOf('master.m3u8') >= 0,
+      'requested only the new playlist'
+  );
+});
+
+QUnit.test('updates the combined segment loader on live playlist refreshes', function() {
+  let updates = [];
+
+  openMediaSource(this.player, this.clock);
+  // master
+  standardXHRResponse(this.requests.shift());
+  // media
+  standardXHRResponse(this.requests.shift());
+
+  this.masterPlaylistController.mainSegmentLoader_.playlist = function(update) {
+    updates.push(update);
+  };
+
+  this.masterPlaylistController.masterPlaylistLoader_.trigger('loadedplaylist');
+  QUnit.equal(updates.length, 1, 'updated the segment list');
+});
+
+QUnit.test(
+'fires a progress event after downloading a segment from combined segment loader',
+function() {
+  let progressCount = 0;
+
+  openMediaSource(this.player, this.clock);
+
+  // master
+  standardXHRResponse(this.requests.shift());
+  // media
+  standardXHRResponse(this.requests.shift());
+
+  this.player.tech_.on('progress', function() {
+    progressCount++;
+  });
+
+  // segment
+  standardXHRResponse(this.requests.shift());
+  this.masterPlaylistController.mainSegmentLoader_.trigger('progress');
+  QUnit.equal(progressCount, 1, 'fired a progress event');
+});
+
+QUnit.test('blacklists switching from video+audio playlists to audio only', function() {
+  let audioPlaylist;
+
+  openMediaSource(this.player, this.clock);
+
+  this.player.tech_.hls.bandwidth = 1e10;
+
+  // master
+  this.requests.shift().respond(200, null,
+                                '#EXTM3U\n' +
+                                '#EXT-X-STREAM-INF:BANDWIDTH=1,CODECS="mp4a.40.2"\n' +
+                                'media.m3u8\n' +
+                                '#EXT-X-STREAM-INF:BANDWIDTH=10,RESOLUTION=1x1\n' +
+                                'media1.m3u8\n');
+  // media1
+  standardXHRResponse(this.requests.shift());
+
+  QUnit.equal(this.masterPlaylistController.masterPlaylistLoader_.media(),
+              this.masterPlaylistController.masterPlaylistLoader_.master.playlists[1],
+              'selected video+audio');
+  audioPlaylist = this.masterPlaylistController.masterPlaylistLoader_.master.playlists[0];
+  QUnit.equal(audioPlaylist.excludeUntil, Infinity, 'excluded incompatible playlist');
+});
+
+QUnit.test('blacklists switching from audio-only playlists to video+audio', function() {
+  let videoAudioPlaylist;
+
+  openMediaSource(this.player, this.clock);
+
+  this.player.tech_.hls.bandwidth = 1;
+  // master
+  this.requests.shift().respond(200, null,
+                                '#EXTM3U\n' +
+                                '#EXT-X-STREAM-INF:BANDWIDTH=1,CODECS="mp4a.40.2"\n' +
+                                'media.m3u8\n' +
+                                '#EXT-X-STREAM-INF:BANDWIDTH=10,RESOLUTION=1x1\n' +
+                                'media1.m3u8\n');
+
+  // media1
+  standardXHRResponse(this.requests.shift());
+  QUnit.equal(this.masterPlaylistController.masterPlaylistLoader_.media(),
+              this.masterPlaylistController.masterPlaylistLoader_.master.playlists[0],
+              'selected audio only');
+  videoAudioPlaylist =
+    this.masterPlaylistController.masterPlaylistLoader_.master.playlists[1];
+  QUnit.equal(videoAudioPlaylist.excludeUntil,
+              Infinity,
+              'excluded incompatible playlist');
+});
+
+QUnit.test('blacklists switching from video-only playlists to video+audio', function() {
+  let videoAudioPlaylist;
+
+  openMediaSource(this.player, this.clock);
+
+  this.player.tech_.hls.bandwidth = 1;
+  // master
+  this.requests.shift()
+    .respond(200, null,
+             '#EXTM3U\n' +
+             '#EXT-X-STREAM-INF:BANDWIDTH=1,CODECS="avc1.4d400d"\n' +
+             'media.m3u8\n' +
+             '#EXT-X-STREAM-INF:BANDWIDTH=10,CODECS="avc1.4d400d,mp4a.40.2"\n' +
+             'media1.m3u8\n');
+
+  // media
+  standardXHRResponse(this.requests.shift());
+  QUnit.equal(this.masterPlaylistController.masterPlaylistLoader_.media(),
+              this.masterPlaylistController.masterPlaylistLoader_.master.playlists[0],
+              'selected video only');
+  videoAudioPlaylist =
+    this.masterPlaylistController.masterPlaylistLoader_.master.playlists[1];
+  QUnit.equal(videoAudioPlaylist.excludeUntil,
+              Infinity,
+              'excluded incompatible playlist');
+});
+
+QUnit.test('blacklists switching between playlists with incompatible audio codecs',
+function() {
+  let alternatePlaylist;
+
+  openMediaSource(this.player, this.clock);
+
+  this.player.tech_.hls.bandwidth = 1;
+  // master
+  this.requests.shift()
+    .respond(200, null,
+             '#EXTM3U\n' +
+             '#EXT-X-STREAM-INF:BANDWIDTH=1,CODECS="avc1.4d400d,mp4a.40.5"\n' +
+             'media.m3u8\n' +
+             '#EXT-X-STREAM-INF:BANDWIDTH=10,CODECS="avc1.4d400d,mp4a.40.2"\n' +
+             'media1.m3u8\n');
+
+  // media
+  standardXHRResponse(this.requests.shift());
+  QUnit.equal(this.masterPlaylistController.masterPlaylistLoader_.media(),
+              this.masterPlaylistController.masterPlaylistLoader_.master.playlists[0],
+              'selected HE-AAC stream');
+  alternatePlaylist =
+    this.masterPlaylistController.masterPlaylistLoader_.master.playlists[1];
+  QUnit.equal(alternatePlaylist.excludeUntil, Infinity, 'excluded incompatible playlist');
+});
+
+QUnit.test('updates the combined segment loader on media changes', function() {
+  let updates = [];
+
+  this.masterPlaylistController.mediaSource.trigger('sourceopen');
+
+  this.masterPlaylistController.hlsHandler.bandwidth = 1;
+
+  // master
+  standardXHRResponse(this.requests.shift());
+  // media
+  standardXHRResponse(this.requests.shift());
+
+  this.masterPlaylistController.mainSegmentLoader_.playlist = function(update) {
+    updates.push(update);
+  };
+
+  // downloading the new segment will update bandwidth and cause a
+  // playlist change
+  // segment 0
+  standardXHRResponse(this.requests.shift());
+  this.masterPlaylistController.mediaSource.sourceBuffers[0].trigger('updateend');
+  // media
+  standardXHRResponse(this.requests.shift());
+  QUnit.equal(updates.length, 1, 'updated the segment list');
+});
+
+QUnit.test('selects a playlist after main/combined segment downloads', function() {
+  let calls = 0;
+
+  this.masterPlaylistController.hlsHandler.selectPlaylist = () => {
+    calls++;
+    return this.masterPlaylistController.masterPlaylistLoader_.master.playlists[0];
+  };
+  this.masterPlaylistController.mediaSource.trigger('sourceopen');
+
+  // master
+  standardXHRResponse(this.requests.shift());
+  // media
+  standardXHRResponse(this.requests.shift());
+
+  // "downloaded" a segment
+  this.masterPlaylistController.mainSegmentLoader_.trigger('progress');
+  QUnit.strictEqual(calls, 2, 'selects after the initial segment');
+
+  // and another
+  this.masterPlaylistController.mainSegmentLoader_.trigger('progress');
+  QUnit.strictEqual(calls, 3, 'selects after additional segments');
+});
+
+QUnit.test('updates the duration after switching playlists', function() {
+  let selectedPlaylist = false;
+
+  this.masterPlaylistController.mediaSource.trigger('sourceopen');
+
+  this.masterPlaylistController.hlsHandler.bandwidth = 1e20;
+
+  // master
+  standardXHRResponse(this.requests[0]);
+  // media
+  standardXHRResponse(this.requests[1]);
+
+  this.masterPlaylistController.hlsHandler.selectPlaylist = () => {
+    selectedPlaylist = true;
+
+    // this duration should be overwritten by the playlist change
+    this.masterPlaylistController.mediaSource = {
+      duration: 0,
+      readyState: 'open'
+    };
+
+    return this.masterPlaylistController.masterPlaylistLoader_.master.playlists[1];
+  };
+
+  // segment 0
+  standardXHRResponse(this.requests[2]);
+  this.masterPlaylistController.mediaSource.sourceBuffers[0].trigger('updateend');
+  // media1
+  standardXHRResponse(this.requests[3]);
+  QUnit.ok(selectedPlaylist, 'selected playlist');
+  QUnit.ok(this.masterPlaylistController.mediaSource.duration !== 0,
+           'updates the duration');
+});
